refactor: hologram demo into models and materials
